pharmaspec-validator/backend/requirements.txt
# FastAPI and async dependencies
fastapi==0.115.13
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database and ORM
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# AI and ML
httpx==0.28.1
ollama==0.4.4
google-generativeai==0.8.3

# Background tasks
celery[redis]==5.4.0
redis==5.2.1

# Data processing
pandas==2.2.3
openpyxl==3.1.5
pypdf==5.1.0

# Environment and config
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Development
black==24.10.0
isort==5.13.2
mypy==1.13.0